refactor(adr-145): split MCP tool definitions by domain
